requirements.txt
fastapi 
uvicorn 
python-multipart
pytesseract
Pillow
pdf2image